+++ R/unitize.core.R
@@ -324,9 +324,11 @@ unitize_core <- function(
 
   global$state("init")  # mark post pre-load state
 
-  # Used to put q/quit here before we switched to tracing them
+  # Maked base functions
 
   util.frame <- new.env(parent=pre.load.frame)
+  assign("quit", unitizer_quit, envir=util.frame)
+  assign("q", unitizer_quit, envir=util.frame)
 
   over_print("Loading unitizer data...")
   eval.which <- seq_along(store.ids)